selectWhere() v2.0 and deleteIndex() fixed

+++ columnas/columnas.c
@@ -191,16 +191,17 @@ TipoRet insertIntoCS(columnas & cs, char *columnasTupla[], char *valoresTupla[],
 }
 
 TipoRet deleteFromCS(columnas & cs, columna c, char *operador, char *valor){
-     if(strcmp(valor, " ") == 0 || (strcasecmp(valor, "EMPTY") == 0 && strcasecmp(calificadorC(c), "NOT_EMPTY") == 0)){
+     if(strcmp(valor, " ") == 0 || ((strcmp(operador, "!") == 0 && strcasecmp(valor, "EMPTY") == 0) && (strcasecmp(calificadorC(c), "NOT_EMPTY") == 0 || strcasecmp(calificadorC(c), "PRIMARY_KEY") == 0))){
           //Se borran todas las tuplas
           cs = deleteAll(cs);
      }else if(strcasecmp(valor, "EMPTY") == 0 && (strcmp(operador, "<") == 0 || strcmp(operador, ">") == 0)){
           //No hago nada
      }else{
           //Debo verificar en la columna si cada valor cumple o no la condicion
-          int index = deleteIndex(c, operador, valor);
-          if(index == 0 && (strcmp(operador, "<") == 0 || strcmp(operador, ">") == 0)){ 
-               //Si el operador es > o < y quiero empezar a eliminar desde el primer elemento, quiero eliminarlos a todos
+          cout<<"intento eliminar  "<< operador <<endl;
+          int index = deleteIndex(c, operador, valor, -1);
+          if(index == 0 && ((strcmp(operador, "<") == 0 || strcmp(operador, ">") == 0) && strcasecmp(calificadorC(c), "PRIMARY_KEY") == 0)){ 
+               //Si el operador es > o < y quiero empezar a eliminar desde el primer elemento de la primary key, quiero eliminarlos a todos
                cs = deleteAll(cs);
           }else{
                while(index != -1){
@@ -209,7 +210,7 @@ TipoRet deleteFromCS(columnas & cs, columna c, char *operador, char *valor){
                          aux->c = deleteFromC(aux->c, index);
                          aux = aux->sig;
                     }
-                    index = deleteIndex(c, operador, valor);
+                    index = deleteIndex(c, operador, valor, index);
                }
           }
      }
@@ -245,10 +246,6 @@ void printdatatableCS(columnas cs){
                     }
                }
           }
-
-
-
-         
      }
 }
 
@@ -271,16 +268,51 @@ void printMetadataCS(columnas cs){
 	}
 }
 
+columnas copiarColumnas(columnas cs){
+     columnas copia = NULL;
+     while(cs!= NULL){
+          addColCS (copia, nombreC(cs->c), tipoDatoC(cs->c), calificadorC(cs->c));
+          cs = cs->sig;
+     }
+     return copia;
+}
 
+columnas copiarDatos(columnas cs1,columna base,columnas cs2, char *valor,char *operador){
+     if(strcmp(valor, " ") == 0 || ((strcmp(operador, "!") == 0 && strcasecmp(valor, "EMPTY") == 0) && (strcasecmp(calificadorC(base), "NOT_EMPTY") == 0 || strcasecmp(calificadorC(base), "PRIMARY_KEY") == 0))){
+          //Copio todas las tuplas
+          cs2 = copiarTodasTuplas(cs1, cs2);
+     }else if(strcasecmp(valor, "EMPTY") == 0 && (strcmp(operador, "<") == 0 || strcmp(operador, ">") == 0)){
+          //No hago nada
+     }else{
+          //Debo verificar en la columna si cada valor cumple o no la condicion
+          int index = deleteIndex(base, operador, valor, -1);
+          if(index == 0 && ((strcmp(operador, "<") == 0 || strcmp(operador, ">") == 0) && strcasecmp(calificadorC(base), "PRIMARY_KEY") == 0)){ 
+               //Si el operador es > o < y quiero empezar a copiar desde el primer elemento de la primary key, quiero copiarlos a todos
+               cs2 = copiarTodasTuplas(cs1, cs2);
+          }else{
+               while(index != -1){
+                    cout<<"intento crear aux = cs2"<<endl;
+                    columnas aux = cs2;
+                    cs1 = revovinarCS(cs1);
+                    while(aux != NULL){
+                         aux->c = copiarValorTupla(cs1->c,aux->c, index);
+                         aux = aux->sig;
+                         if(cs1->sig != NULL)
+                              cs1 = cs1->sig;
+                    }
+                    index = deleteIndex(base, operador, valor, index);
+               }
+          }
+     }
+     return cs2;
+}
 
-
-
-
-
-
-
-
-
-
-
-
+columnas copiarTodasTuplas(columnas cs1,columnas cs2){
+     columnas aux = cs2;
+     while(aux != NULL){
+          aux->c = copiarTodasTuplasC(cs1->c, aux->c);
+          aux = aux->sig;
+          cs1 = cs1->sig;
+     }
+	return cs2;
+}
